pyproject.toml
[tool.poetry]
name = "deplatformer-webapp"
version = "0.1.7"
description = "Flask webapp for the deplatformer platform"
authors = ["Deplatformer Team <info@deplatformer.io>"]
maintainers = ["Antreas Pogiatzis <antreas@deplatformer.io>"]
license = "MIT"
readme = "README.md"
homepage = "https://deplatformer.io"
repository = "https://github.com/deplatformer/deplatformer-webapp"
keywords = [
    "Filecoin",
    "IPFS",
    "web3",
    "Deplatformer"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Flake8",
    "Framework :: Flask",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Topic :: Communications :: File Sharing",
    "Topic :: Internet",
    "Topic :: Multimedia",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
include = ["deplatformer_webapp/migrations/*"]
exclude = [".DS_Store"]


[tool.poetry.dependencies]
python = "^3.7.2"
bcrypt = "3.2.0"
blinker = "1.4"
cffi = "1.14.2"
click = "7.1.2"
cryptography = "3.1"
Deprecated = "1.2.10"
dnspython = "1.16.0"
email-validator = "1.1.1"
eventlet = "0.27.0"
Flask = "1.1.2"
Flask-Login = "0.5.0"
Flask-Mail = "0.9.1"
Flask-SocketIO = "4.3.1"
Flask-SQLAlchemy = "2.4.4"
Flask-User = "^1.0.2"
Flask-WTF = "0.14.3"
ftfy = "5.8"
greenlet = "0.4.16"
grpcio = "1.30.0"
idna = "^2.10"
itsdangerous = "1.1.0"
Jinja2 = "2.11.2"
MarkupSafe = "1.1.1"
monotonic = "1.5"
urllib3= "1.25.11"
passlib = "1.7.2"
protobuf = "3.12.4"
pycparser = "2.20"
pygate-grpc = "0.0.11"
python-engineio = "3.13.2"
python-socketio = "4.6.0"
six = "1.15.0"
SQLAlchemy = "1.3.19"
wcwidth = "0.2.5"
Werkzeug = "1.0.1"
wrapt = "1.12.1"
WTForms = "2.3.3"
poetry-version = "^0.1.5"
Flask-Migrate = "^2.5.3"
python-dotenv = "^0.15.0"
docker = "^4.3.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
pytest = "^6.1.2"
isort = "^5.6.4"
pylint = "^2.6.0"
flakehell = "^0.7.0"

[tool.poetry.scripts]
deplatformer = 'deplatformer_webapp.cli:cli'
format = 'deplatformer_webapp.scripts:format'
lint = 'deplatformer_webapp.scripts:lint'
dropdb = 'deplatformer_webapp.scripts:dropdb'

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.tox
    | \.venv
    | migrations
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.flakehell]
exclude = ["README.md", "node_modules/"]
format = "grouped"
max_line_length = 120
show_source = true

[tool.flakehell.plugins]
flake8-bandit = ["+*", "-S322"]
flake8-bugbear = ["+*"]
flake8-builtins = ["+*"]
flake8-comprehensions = ["+*"]
flake8-darglint = ["+*"]
flake8-docstrings = ["+*"]
flake8-eradicate = ["+*"]
flake8-isort = ["+*"]
flake8-mutable = ["+*"]
flake8-pytest-style = ["+*"]
flake8-spellcheck = ["+*"]
mccabe = ["+*"]
pep8-naming = ["+*"]
pycodestyle = ["+*"]
pyflakes = ["+*"]
pylint = ["+*"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
